rectification des controllers users et posts et du middleware auth

+++ controllers/users.js
@@ -1,7 +1,6 @@
 const User = require("../models/User");
 const fs = require('fs');
 
-
 //Logique de récupération d'un seul utilisateur
 exports.getOneUser = (req, res) => {
     User.findOne({ _id: req.params.id })
@@ -11,7 +10,7 @@ exports.getOneUser = (req, res) => {
             }
             res.status(200).json(user);
         })
-        .catch(error => res.status(500).json({ error }));
+        .catch(error => res.status(500).json({ message: error.message }));
 };
 
 //Logique de la mise à jour d'un utilisateur et/ou de son image
@@ -26,19 +25,22 @@ exports.updateOneUser = (req, res) => {
             if (!user) {
                 return res.status(404).json({ message: "utilisateur non trouvé" })
             }
-            const oldImageName = user.imageUrl.split('/images/users/')[1];
+            let oldImageName;
+            if (user.imageUrl) {
+                oldImageName = user.imageUrl.split('/images/users/')[1];
+            }
             user.updateOne({ _id: req.params.id }, { ...userObject })
                 .then(() => {
-                    if (req.file) {
+                    if (req.file && oldImageName) {
                         fs.unlink(`images/${oldImageName}`, () => {
                             console.log("fichier supprimé");
                         });
                     }
                     res.status(200).json({ message: "mis à jour de l'utilisateur" })
                 })
-                .catch(error => res.status(400).json({ error }));
+                .catch(error => res.status(400).json({ message: error.message }));
         })
-        .catch(error => res.status(500).json({ error }));
+        .catch(error => res.status(500).json({ message: error.message }));
 };
 
 //Logique de la suppression d'un utilisateur et son image
@@ -48,15 +50,20 @@ exports.deleteOneUser = (req, res) => {
             if (!user) {
                 return res.status(404).json({ message: "Utilisateur non trouvé" })
             }
-            const filename = user.imageUrl.split('/images/users/')[1];
+            let oldImageName;
+            if (user.imageUrl) {
+                oldImageName = user.imageUrl.split('/images/users/')[1];
+            }
             User.deleteOne({ _id: req.params.id })
                 .then(() => {
-                    fs.unlink(`images/users/${filename}`, () => {
-                        console.log("fichier supprimé");
-                    });
+                    if (oldImageName) {
+                        fs.unlink(`images/users/${filename}`, () => {
+                            console.log("fichier supprimé");
+                        });
+                    }
                     res.status(200).json({ message: 'Utilisateur supprimé' })
                 })
-                .catch(error => res.status(400).json({ error }));
+                .catch(error => res.status(400).json({ message: error.message }));
         })
-        .catch(error => res.status(500).json({ error }));
+        .catch(error => res.status(500).json({ message: error.message }));
 };
